+++ examples/versions.tf
@@ -7,9 +7,9 @@ terraform {
 
     ct = {
       source  = "poseidon/ct"
-      version = "~> 0.7.1"
+      version = "~> 0.8.0"
       #source = "terraform.localhost/poseidon/ct"
-      #version = "0.7.1"
+      #version = "0.8.0"
     }
   }
 }